:construction: split the builder into multiple components
